@@ -3,58 +3,61 @@
   "version": "1.0.5",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --port=8081 -o=true",
-    "build": "ng build --configuration=production",
+    "start": "ng serve -o",
+    "build": "ng build --configuration production",
     "watch": "ng build --watch --configuration development",
-    "test": "ng test"
+    "test": "ng test",
+    "run:all": "node node_modules/@angular-architects/module-federation/src/server/mf-dev-server.js"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "16.1.0",
-    "@angular/common": "16.1.0",
-    "@angular/compiler": "16.1.0",
-    "@angular/core": "16.1.0",
-    "@angular/forms": "16.1.0",
-    "@angular/platform-browser": "16.1.0",
-    "@angular/platform-browser-dynamic": "16.1.0",
-    "@angular/router": "16.1.0",
+    "@angular-architects/module-federation": "17.0.8",
+    "@angular/animations": "17.2.4",
+    "@angular/common": "17.2.4",
+    "@angular/compiler": "17.2.4",
+    "@angular/core": "17.2.4",
+    "@angular/forms": "17.2.4",
+    "@angular/platform-browser": "17.2.4",
+    "@angular/platform-browser-dynamic": "17.2.4",
+    "@angular/router": "17.2.4",
     "@ngx-translate/core": "15.0.0",
     "@ngx-translate/http-loader": "8.0.0",
     "angular-svg-icon": "16.0.0",
     "device-uuid": "1.0.4",
     "firebase": "10.5.2",
     "lodash": "4.17.21",
-    "ng-zorro-antd": "16.1.0",
+    "ng-zorro-antd": "17.2.0",
     "rxjs": "7.8.0",
     "tslib": "2.3.0",
-    "zone.js": "0.13.0"
+    "zone.js": "0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "16.1.4",
-    "@angular-eslint/builder": "14.1.2",
-    "@angular-eslint/eslint-plugin": "14.1.2",
-    "@angular-eslint/eslint-plugin-template": "14.1.2",
-    "@angular-eslint/schematics": "14.1.2",
-    "@angular-eslint/template-parser": "14.1.2",
-    "@angular/cli": "16.1.4",
-    "@angular/compiler-cli": "16.1.0",
+    "@angular-devkit/build-angular": "17.2.3",
+    "@angular-eslint/builder": "17.2.1",
+    "@angular-eslint/eslint-plugin": "17.2.1",
+    "@angular-eslint/eslint-plugin-template": "17.2.1",
+    "@angular-eslint/schematics": "17.2.1",
+    "@angular-eslint/template-parser": "17.2.1",
+    "@angular/cli": "17.2.3",
+    "@angular/compiler-cli": "17.2.4",
     "@types/device-uuid": "1.0.3",
     "@types/jasmine": "4.3.0",
     "@types/lodash": "4.14.197",
     "@types/node": "20.8.8",
-    "@typescript-eslint/eslint-plugin": "5.37.0",
-    "@typescript-eslint/parser": "5.37.0",
+    "@typescript-eslint/eslint-plugin": "^6.10.0",
+    "@typescript-eslint/parser": "^6.10.0",
     "autoprefixer": "10.4.15",
-    "eslint": "8.47.0",
+    "eslint": "^8.53.0",
     "jasmine-core": "4.6.0",
     "karma": "6.4.0",
     "karma-chrome-launcher": "3.2.0",
     "karma-coverage": "2.2.0",
     "karma-jasmine": "5.1.0",
     "karma-jasmine-html-reporter": "2.1.0",
+    "ngx-build-plus": "^17.0.0",
     "postcss": "8.4.29",
     "tailwindcss": "3.3.5",
     "ts-loader": "9.5.1",
-    "typescript": "5.1.3"
+    "typescript": "5.3.3"
   }
 }
